--- bin/update-go
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-
-
-exit 0
