+++ cmd/gen-release-notes/main.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/rancher/ecm-distro-tools/pkg/repository"
+	"github.com/rancher/ecm-distro-tools/repository"
 )
 
 var (
